Add confirmation popup before removing

+++ app/assets/javascripts/index/index.coffee
@@ -1,10 +1,11 @@
 class @Index
   @delete = (path) ->
-    $.ajax path,
-      type: 'DELETE'
-      dataType: 'html'
-      error: (jqXHR, textStatus, errorThrown) ->
-        $('body').append "AJAX Error: #{textStatus}"
-      success: (data, textStatus, jqXHR) ->
-        $('body').append "Successful AJAX call: #{data}"
+    if confirm "Do you really want to delete `#{path}'?"
+      $.ajax path,
+        type: 'DELETE'
+        dataType: 'html'
+        error: (jqXHR, textStatus, errorThrown) ->
+          $('body').append "AJAX Error: #{textStatus}"
+        success: (data, textStatus, jqXHR) ->
+          $('body').append "Successful AJAX call: #{data}"
     false
